feat: delete sessions only if location is commcare
Former-commit-id: e79da66b7bce8db2309e08eba444722939c78cf7 [formerly 9cfb3f6385ce18aa423354a53e47c72a2791dbdf]
Former-commit-id: b3c9654cbee41bacf2ec14f02ee2b7f2223a3dfe

+++ clientapi/src/main/java/com/simprints/clientapi/activities/commcare/CommCarePresenter.kt
@@ -66,7 +66,7 @@ class CommCarePresenter(
                 flowCompletedCheck,
                 getEventsJsonForSession(getCurrentSessionIdOrEmpty())
             )
-            deleteSessionEvents(getCurrentSessionIdOrEmpty())
+            deleteSessionEventsIfNeeded(getCurrentSessionIdOrEmpty())
         }
     }
 
@@ -98,7 +98,7 @@ class CommCarePresenter(
                 getCurrentSessionIdOrEmpty(),
                 getEventsJsonForSession(getCurrentSessionIdOrEmpty())
             )
-            deleteSessionEvents(getCurrentSessionIdOrEmpty())
+            deleteSessionEventsIfNeeded(getCurrentSessionIdOrEmpty())
         }
     }
 
@@ -112,7 +112,7 @@ class CommCarePresenter(
                 getCurrentSessionIdOrEmpty(),
                 getEventsJsonForSession(getCurrentSessionIdOrEmpty())
             )
-            deleteSessionEvents(getCurrentSessionIdOrEmpty())
+            deleteSessionEventsIfNeeded(getCurrentSessionIdOrEmpty())
         }
     }
 
@@ -128,7 +128,7 @@ class CommCarePresenter(
                 flowCompletedCheck,
                 getEventsJsonForSession(getCurrentSessionIdOrEmpty())
             )
-            deleteSessionEvents(getCurrentSessionIdOrEmpty())
+            deleteSessionEventsIfNeeded(getCurrentSessionIdOrEmpty())
         }
     }
 
@@ -143,7 +143,7 @@ class CommCarePresenter(
                 flowCompletedCheck,
                 getEventsJsonForSession(getCurrentSessionIdOrEmpty())
             )
-            deleteSessionEvents(getCurrentSessionIdOrEmpty())
+            deleteSessionEventsIfNeeded(getCurrentSessionIdOrEmpty())
         }
     }
 
@@ -170,8 +170,10 @@ class CommCarePresenter(
         processConfirmIdentityRequest()
     }
 
-    private suspend fun deleteSessionEvents(sessionId: String) {
-        sessionEventsManager.deleteSessionEvents(sessionId)
+    private suspend fun deleteSessionEventsIfNeeded(sessionId: String) {
+        if (sharedPreferencesManager.syncDestinationSetting == SyncDestinationSetting.COMMCARE) {
+            sessionEventsManager.deleteSessionEvents(sessionId)
+        }
     }
 
     private data class CommCareEvents(val events: List<Event>)
